refactor: extract IssueScreenshotAnnotation.swift from IssueCoreSubtypes.swift
Splits IssueCoreSubtypes.swift by concern:
- IssueCoreSubtypes.swift keeps category / severity enums +
  IssueReproductionStep
- IssueScreenshotAnnotation.swift gets the screenshot-annotation
  model (including its nested Style enum, ~93 lines)

Byte-identical move of lines 54-147. Both files import Foundation.

IssueCoreSubtypes.swift: 147 → 52 lines. Tests: 667.
